v0.4.184: split AssignmentRoutes+Editor.swift into four cohesive files
Phase 4.2 of the audit-driven refactor.  Counterpart to v0.4.183
(Phase 4.1, ManifestValidation split).

`AssignmentRoutes+Editor.swift` (809 LOC) mixed four conceptually
distinct handler groups under a single `extension
PublishedAssignmentRoutes`.  Split into four files, each under 330
LOC, with names that match their new parent type (Phase 2 deferred
the file rename from `AssignmentRoutes+*` to `PublishedAssignmentRoutes+*`
for blame continuity; this PR does it for the four split files):

  - `PublishedAssignmentRoutes+FileDownloads.swift` (109 LOC) — the
    three `GET /instructor/:assignmentID/files/...` endpoints.
  - `PublishedAssignmentRoutes+SaveEdit.swift` (328 LOC) —
    `saveEditedAssignment` plus its seven file-private helpers.
  - `PublishedAssignmentRoutes+ScriptCRUD.swift` (281 LOC) — the
    four `:assignmentID/scripts` endpoints plus the
    `safeScriptFilename(from:)` free helper.
  - `PublishedAssignmentRoutes+NotebookTools.swift` (140 LOC) —
    `script-templates`, `scan-notebook`, and `create-solution`.

Also updated the in-scope file list in
`WebAssignmentErrorTests.noRawAbortInInstructorAssignmentRoutes` to
reference the four new filenames.

No behaviour change.  1279/1279 tests pass; 0 SwiftLint violations.
